Rename config directory so it's not used implicitly
Spring-boot searches for application property files in the /config directory
of the current directory [1]. This means that when running the example client
from the project root, it would be picking up the /config directory directly
(rather than only via the config server). This wasn't intended, so the
directory is being renamed.

[1] http://docs.spring.io/spring-boot/docs/current/reference/htmlsingle/#boot-features-external-config-application-property-files